requirements.txt
google-generativeai>=0.7.2,<0.8
python-dotenv>=1.0.1
pandas>=2.0.0
numpy>=1.24.0
streamlit>=1.31.0
pinecone>=5.0.0
